Added support for sub-second resolution

+++ src/main/java/org/samcrow/antrecorder/TimedEvent.java
@@ -13,11 +13,11 @@ public class TimedEvent {
     /**
      * The time of the event, in seconds
      */
-    private final int time;
+    private final double time;
     
     private final Event event;
 
-    public TimedEvent(int time, Event event) {
+    public TimedEvent(double time, Event event) {
         this.time = time;
         this.event = event;
     }
@@ -26,7 +26,7 @@ public class TimedEvent {
      * 
      * @return The time of this event, in seconds
      */
-    public int getTime() {
+    public double getTime() {
         return time;
     }
